backend/requirements.txt
Django>=5.0,<5.3
djangorestframework>=3.15.0
django-cors-headers>=4.4.0
djangorestframework-simplejwt>=5.3.1
drf-yasg>=1.21.8
python-dotenv>=1.0.1
neo4j>=5.25.0
openai>=1.52.0
psycopg2-binary>=2.9.10
Pillow>=10.4.0
gunicorn>=21.2.0
requests>=2.32.0
